DiscordFix/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.1",
    "discord-py>=2.6.3",
    "motor>=3.7.1",
    "pydantic>=2.11.9",
    "pymongo>=4.15.1",
    "python-dotenv>=1.1.1",
]
